[BEAM-882,BEAM-878] Simplified creating new API surface verifications.
