+++ sdks/java/core/src/test/java/org/apache/beam/sdk/util/ApiSurfaceTest.java
@@ -17,123 +17,22 @@
  */
 package org.apache.beam.sdk.util;
 
-import static org.hamcrest.Matchers.anyOf;
 import static org.hamcrest.Matchers.containsInAnyOrder;
 import static org.hamcrest.Matchers.emptyIterable;
 import static org.junit.Assert.assertThat;
-import static org.junit.Assert.fail;
 
-import com.google.common.base.Joiner;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Lists;
-import com.google.common.collect.Maps;
 import com.google.common.collect.Sets;
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import org.hamcrest.Description;
-import org.hamcrest.Matcher;
-import org.hamcrest.TypeSafeDiagnosingMatcher;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
 
 /**
- * Tests for ApiSurface. These both test the functionality and also that our
- * public API is conformant to a hard-coded policy.
+ * Functionality tests for ApiSurface.
  */
 @RunWith(JUnit4.class)
 public class ApiSurfaceTest {
-
-  @Test
-  public void testOurApiSurface() throws Exception {
-    ApiSurface checkedApiSurface = ApiSurface.getSdkApiSurface();
-
-    Map<Class<?>, List<Class<?>>> disallowedClasses = Maps.newHashMap();
-    for (Class<?> clazz : checkedApiSurface.getExposedClasses()) {
-      if (!classIsAllowed(clazz)) {
-        disallowedClasses.put(clazz, checkedApiSurface.getAnyExposurePath(clazz));
-      }
-    }
-
-    List<String> disallowedMessages = Lists.newArrayList();
-    for (Map.Entry<Class<?>, List<Class<?>>> entry : disallowedClasses.entrySet()) {
-      disallowedMessages.add(entry.getKey() + " exposed via:\n\t\t"
-      + Joiner.on("\n\t\t").join(entry.getValue()));
-    }
-    Collections.sort(disallowedMessages);
-
-    if (!disallowedMessages.isEmpty()) {
-      fail("The following disallowed classes appear in the public API surface of the SDK:\n\t"
-        + Joiner.on("\n\t").join(disallowedMessages));
-    }
-  }
-
-  @SuppressWarnings("unchecked")
-  private static final Set<Matcher<Class<?>>> ALLOWED_PACKAGES =
-      ImmutableSet.of(
-          inPackage("org.apache.beam"),
-          inPackage("com.google.api.client"),
-          inPackage("com.google.api.services.bigquery"),
-          inPackage("com.google.api.services.cloudresourcemanager"),
-          inPackage("com.google.api.services.dataflow"),
-          inPackage("com.google.api.services.pubsub"),
-          inPackage("com.google.api.services.storage"),
-          inPackage("com.google.auth"),
-          inPackage("com.google.bigtable.v1"),
-          inPackage("com.google.cloud.bigtable.config"),
-          inPackage("com.google.cloud.bigtable.grpc"),
-          inPackage("com.google.datastore"),
-          inPackage("com.google.protobuf"),
-          inPackage("com.google.rpc"),
-          inPackage("com.google.type"),
-          inPackage("com.fasterxml.jackson.annotation"),
-          inPackage("com.fasterxml.jackson.core"),
-          inPackage("com.fasterxml.jackson.databind"),
-          inPackage("com.fasterxml.jackson.deser"),
-          inPackage("io.grpc"),
-          inPackage("org.apache.avro"),
-          inPackage("org.apache.commons.logging"), // via BigTable
-          inPackage("org.hamcrest"), // via DataflowMatchers
-          inPackage("org.codehaus.jackson"), // via Avro
-          inPackage("org.joda.time"),
-          inPackage("org.junit"),
-          inPackage("java"));
-
-  @SuppressWarnings({"rawtypes", "unchecked"})
-  private boolean classIsAllowed(Class<?> clazz) {
-    // Safe cast inexpressible in Java without rawtypes
-    return anyOf((Iterable) ALLOWED_PACKAGES).matches(clazz);
-  }
-
-  private static Matcher<Class<?>> inPackage(String packageName) {
-    return new ClassInPackage(packageName);
-  }
-
-  private static class ClassInPackage extends TypeSafeDiagnosingMatcher<Class<?>> {
-
-    private final String packageName;
-
-    public ClassInPackage(String packageName) {
-      this.packageName = packageName;
-    }
-
-    @Override
-    public void describeTo(Description description) {
-      description.appendText("Class in package \"");
-      description.appendText(packageName);
-      description.appendText("\"");
-    }
-
-    @Override
-    protected boolean matchesSafely(Class<?> clazz, Description mismatchDescription) {
-      return clazz.getName().startsWith(packageName + ".");
-    }
-
-  }
-
-  //////////////////////////////////////////////////////////////////////////////////
 
   @SuppressWarnings({"rawtypes", "unchecked"})
   private void assertExposed(Class classToExamine, Class... exposedClasses) {
